+++ framework/include/ImageWriter/CoreAndAPIShared.h
@@ -6,8 +6,8 @@
 
 struct InstanceData;
 typedef bool (*PfnCoreInitialize)(InstanceData*);
-typedef bool (*PfnCoreDrawCircle)(InstanceData, int, int, int);
-typedef bool (*PfnCoreDrawRectangle)(InstanceData, int, int, int, int);
+typedef bool (*PfnCoreDrawCircle)(InstanceData, float, float, float);
+typedef bool (*PfnCoreDrawRectangle)(InstanceData, float, float, float, float);
 typedef bool (*PfnCoreClearImage)(InstanceData);
 typedef bool (*PfnCoreInitRenderWindow)(InstanceData, int64_t);
 typedef bool (*PfnCoreUpdateRenderWindow)(InstanceData, float);
